types refactored
