+++ src/nfs/io/reader.rs
@@ -18,21 +18,18 @@ use nfs;
 use std::sync;
 use super::network_storage::NetworkStorage;
 use self_encryption;
-use std::fs;
-use rand;
-use rand::Rng;
 use routing;
 use client;
 
 pub struct Reader {
-    file: nfs::types::File,
+    file: nfs::file::File,
     self_encryptor: self_encryption::SelfEncryptor<NetworkStorage>,
     client: ::std::sync::Arc<::std::sync::Mutex<client::Client>>
 }
 
 impl Reader {
 
-    pub fn new(file: nfs::types::File,
+    pub fn new(file: nfs::file::File,
         client: ::std::sync::Arc<::std::sync::Mutex<client::Client>>) -> Reader {
         let storage = sync::Arc::new(NetworkStorage::new(client.clone()));
         Reader {
