+++ src/nfs/mod.rs
@@ -15,6 +15,10 @@
 // Please review the Licences for the specific language governing permissions and limitations
 // relating to use of the SAFE Network Software.
 
-mod types;
 mod io;
 mod helper;
+mod traits;
+mod file;
+mod metadata;
+mod directory_info;
+mod directory_listing;
